refactor: 패캐지명 오타 변경
contant -> constant 로 변경
